Apply the configured keyboard layout on the SDDM greeter
The Wayland greeter runs its own minimal Hyprland from default/sddm/hyprland.lua,
which only set misc/animations. The layout resolution from /etc/vconsole.conf
lived solely in default/hypr/input.lua, so the greeter always fell back to
us/QWERTY -- users with a non-us layout could not type their password the way
they set it.

Extract the vconsole -> kb_layout resolution into default/hypr/keyboard.lua and
have both the session (input.lua) and the greeter share it, so the password
field uses the same layout as the session. The greeter loads the module
defensively: any failure falls back to Hyprland's default rather than breaking
the login screen.
