+++ test/shell.d/hyprland-keyboard-layout-test.sh
@@ -74,12 +74,72 @@ assert_input "non-latin layout in front gains us even when us trails" "[us,il,us
 '
 
 hooks_conf="$ROOT/etc/mkinitcpio.conf.d/omarchy_hooks.conf"
-input_lua="$ROOT/default/hypr/input.lua"
+keyboard_lua="$ROOT/default/hypr/keyboard.lua"
 
 hooks_layouts=$(awk -F')' '/\) ;;$/ { gsub(/[[:space:]|]+/, "\n", $1); print $1 }' "$hooks_conf" | grep '^[a-z]\+$' | sort)
-lua_layouts=$(sed -n '/^local non_latin_layouts =/,+1p' "$input_lua" | grep -o '"[^"]*"' | tr -d '"' | tr ' ' '\n' | grep '^[a-z]\+$' | sort)
+lua_layouts=$(sed -n '/^local non_latin_layouts =/,+1p' "$keyboard_lua" | grep -o '"[^"]*"' | tr -d '"' | tr ' ' '\n' | grep '^[a-z]\+$' | sort)
 
 [[ -n $hooks_layouts ]] || fail "non-latin layout list is readable from omarchy_hooks.conf"
 [[ $hooks_layouts == "$lua_layouts" ]] ||
   fail "non-latin layout lists stay in sync" "$(diff <(echo "$hooks_layouts") <(echo "$lua_layouts"))"
 pass "non-latin layout lists stay in sync with the initramfs hook"
+
+# The SDDM greeter runs its own minimal Hyprland without bootstrap.lua, so it
+# must reach the shared resolver on its own and hand the greeter the same layout
+# the session uses. Otherwise the password field falls back to us/QWERTY.
+greeter_input() {
+  OMARCHY_PATH="$ROOT" OMARCHY_VCONSOLE="${1-}" lua <<'LUA'
+local vconsole = os.getenv("OMARCHY_VCONSOLE")
+local real_open = io.open
+
+io.open = function(path, mode)
+  if path ~= "/etc/vconsole.conf" then
+    return real_open(path, mode)
+  end
+
+  if not vconsole then
+    return nil
+  end
+
+  local file = io.tmpfile()
+  file:write(vconsole)
+  file:seek("set")
+  return file
+end
+
+hl = {
+  config = function(config)
+    local input = config.input
+    print(("[%s] [%s] [%s]"):format(
+      input.kb_layout or "nil",
+      input.kb_variant or "nil",
+      input.kb_options or "nil"
+    ))
+  end,
+}
+
+require("default.sddm.hyprland")
+LUA
+}
+
+assert_greeter() {
+  local description="$1"
+  local expected="$2"
+  local actual
+
+  if (( $# > 2 )); then
+    actual=$(greeter_input "$3")
+  else
+    actual=$(greeter_input)
+  fi
+
+  [[ $actual == "$expected" ]] ||
+    fail "$description" "expected: $expected"$'\n'"actual:   $actual"
+  pass "$description"
+}
+
+assert_greeter "greeter types the configured latin layout" "[fr] [] [$base_options]" 'XKBLAYOUT=fr
+'
+assert_greeter "greeter mirrors the session for non-latin layouts" "[us,ru] [,] [$toggle_options]" 'XKBLAYOUT=ru
+'
+assert_greeter "greeter falls back to us without vconsole" "[us] [] [$base_options]"
